When getting the serial port buffer size, use a vpr::Uint16 to allow for larger buffer sizes.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@4940 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ juggler/modules/vapor/vpr/IO/Port/SerialPort_t.h
@@ -238,13 +238,13 @@ public:
     //! POST: The maximum buffer size is returned to the caller through the
     //+       by-reference argument.
     //
-    //! ARGS: size - A reference to a vpr::Uint8 where the buffer size is
+    //! ARGS: size - A reference to a vpr::Uint16 where the buffer size is
     //+              stored for return to the caller.
     //
     //! RETURNS: A vpr::Status object describing the results of the operation.
     // ------------------------------------------------------------------------
     inline vpr::Status
-    getBufferSize (vpr::Uint8& size) {
+    getBufferSize (vpr::Uint16& size) {
         return m_sio_imp.getBufferSize(size);
     }
